Properly fill in display_width and display_height for 0 now.

+++ src/components/TextEditComponent.cpp
@@ -73,7 +73,7 @@ void TextEditComponent::onTextChanged()
 	{
 		float y = getFont()->sizeWrappedText(mText, mSize.x()).y();
 		if(y == 0)
-			y = getFont()->getHeight();
+			y = (float)getFont()->getHeight();
 		
 		setSize(mSize.x(), y);
 	}
